Consolidate `QueryActionContextProvider` utils
